feat: refactor Juego model and service to use JuegoTypes, add visitas field
