+++ src/types/juego.types.ts
@@ -1,3 +1,27 @@
+import { Types } from "mongoose";
+
+export interface IJuego extends Document {
+  titulo: string;
+  descripcion: string;
+  tutorial: string;
+  cant_min_pers: number;
+  cant_max_pers: number;
+  duracion_max: number;
+  duracion_min: number;
+  precio: number;
+  imagen: string;
+  disponible: boolean;
+  activo: boolean;
+  justificacionRetiro?: string 
+  cantidad: number;
+  cantidad_prestamo: number;
+  id_dificultad: Types.ObjectId;
+  id_editorial: Types.ObjectId;
+  createdAt?: Date;
+  updatedAt?: Date;
+  visitas: number
+}
+
 export interface GameNumericItems {
   cant_min_pers: number;
   cant_max_pers: number;
